Bug fix for banned classes getting weird randoms
So the previous logic was assuming that if the character's class was in the `promoted` list, then set to `promoted`, otherwise set to `unpromoted`. Well, if the class is banned like the Demon King's, it'll fall under `unpromoted`. So I've created a banned promoted list so we can set banned characters to promotes this way, or unpromotes the other way.