Give Legault a chest key
It turns out that Legault's AI is set so that he will try to unlock all chests and then immediately leave. Well in Eliwood's journey, he starts with a lockpick so if he's not a thief, he has no way of unlocking chests and will immediately leave on turn 1. Give him a chest key and he'll happily continue on until he uses it up.