Fix several item and char randomization bugs
There's a place where Gale and Milady try to leave the chapter,
but it causes the game to freeze because they try to walk across
mountains. I can't figure out where to control that movement
starting and ending position so I'm hoping that Gale will just be
randomized in the chapter you fight him, but not in this chapter.