Fix bug where Pascal's and Kenneth's weapon levels were wrong
So this was all due to a copy/paste error on my part. When copying the
character blocks, I forgot to change the IDs on Mark and the Nils and
Ninian children characters from chapther 19x part 2. So I updated the
IDs and ended up removing one of the children characters because it
shared the same character ID as Jan. So I guess both Jan and this child
will always have the same class.